refactor: move RTC commands into cli_commands/rtc.py
